pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Erasyl",email = "erasylbidan@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (==0.109.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
